requirements.txt
streamlit==1.41.0
pandas==1.5.3
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.1
numpy==1.24.0
xgboost==1.7.6
streamlit-reveal-slides==0.1.0
